requirements.txt
pandas~=1.4.3
prefect~=2.7.10
prefect-sqlalchemy==0.2.2
prefect-gcp[cloud_storage]==0.2.4
protobuf~=4.21.12
pyarrow~=11.0.0
pandas-gbq==0.18.1
psycopg2-binary==2.9.5
sqlalchemy~=1.4.9
h11~=0.12.0
jupyter~=1.0.0
pip~=22.3.1
attrs~=20.3.0
cryptography~=36.0.2
docutils~=0.16
tornado~=6.1
Jinja2~=3.0.3
rsa~=4.7.2
pyasn1~=0.4.8
Mako~=1.2.0
Babel~=2.10.1
Pygments~=2.12.0
MarkupSafe~=2.0.1
pytz~=2022.1
rich~=12.4.4
PyYAML~=5.4.1
anyio~=3.6.1
sniffio~=1.2.0
click~=8.1.3
httpcore~=0.15.0
idna~=3.3
rfc3986~=1.5.0
certifi~=2020.12.5
pypiwin32~=223
pywin32~=304
numpy~=1.23.0
setuptools~=58.1.0
cffi~=1.15.0
requests~=2.27.1
packaging~=21.3
lz4~=4.3.2
six~=1.16.0
urllib3~=1.26.9
ipython~=8.4.0
python-dateutil~=2.8.2
colorama~=0.4.4
alembic~=1.8.0
Markdown~=3.3.7
pendulum~=2.1.2
httpx~=0.23.0
croniter~=1.3.5
pathspec~=0.9.0
tzdata~=2022.7
gunicorn~=20.1.0
greenlet~=1.1.2
psutil~=5.9.1
pyparsing~=2.4.7
blinker~=1.4
pytzdata~=2020.1
psycopg2~=2.9.3
decorator~=5.1.1
pycparser~=2.21
itsdangerous~=1.1.0
wget~=3.2
hpack~=4.0.0
hyperframe~=6.0.1
h2~=4.1.0
toml~=0.10.2
regex~=2022.10.31
typer~=0.7.0
websocket-client~=1.5.0
docker~=6.0.1
paramiko~=3.0.0
fsspec~=2023.1.0
cachetools~=5.3.0
google-api-python-client~=2.75.0
storage~=0.0.4.3
griffe~=0.25.4
orjson~=3.8.5
apprise~=1.2.1
asyncpg~=0.27.0
fastapi~=0.89.1
starlette~=0.22.0
pydantic~=1.10.4
readchar~=4.0.3
uvicorn~=0.20.0
dateparser~=1.1.6
kubernetes~=25.3.0
cloudpickle~=2.2.1
coolname~=2.2.0
jsonpatch~=1.32
tzlocal~=4.2
httplib2~=0.21.0
oauthlib~=3.2.2
aiosqlite~=0.18.0
uritemplate~=4.1.1
bcrypt~=4.0.1
jsonpointer~=2.3